+++ saleor/checkout/migrations/0040_add_handle_checkouts_permission.py
@@ -43,9 +43,9 @@ def assign_permissions(apps, schema_editor):
 
 class Migration(migrations.Migration):
     dependencies = [
+        ("checkout", "0039_alter_checkout_email"),
         ("product", "0159_auto_20220209_1501"),
         ("order", "0133_rename_order_token_id"),
-        ("checkout", "0039_alter_checkout_email"),
     ]
 
     operations = [