+++ saleor/graphql/core/tests/test_file_validation.py
@@ -11,34 +11,9 @@ from ....product.error_codes import ProductErrorCode
 from ..validators.file import (
     clean_image_file,
     detect_mime_type,
-    get_mime_type,
-    is_image_mimetype,
     is_supported_image_mimetype,
-    is_valid_image_content_type,
     validate_upload_file,
 )
-
-
-def test_is_image_mimetype_valid_mimetype():
-    # given
-    valid_mimetype = "image/jpeg"
-
-    # when
-    result = is_image_mimetype(valid_mimetype)
-
-    # then
-    assert result
-
-
-def test_is_image_mimetype_invalid_mimetype():
-    # given
-    invalid_mimetype = "application/javascript"
-
-    # when
-    result = is_image_mimetype(invalid_mimetype)
-
-    # then
-    assert not result
 
 
 def test_is_supported_image_mimetype_valid_mimetype():
@@ -61,31 +36,6 @@ def test_is_supported_image_mimetype_invalid_mimetype():
 
     # then
     assert not result
-
-
-@pytest.mark.parametrize(
-    ("content_type", "is_valid"),
-    [
-        ("image/jpeg", True),
-        ("image/jpg", True),
-        ("image/png", True),
-        ("image/gif", True),
-        ("image/bmp", True),
-        ("image/tiff", True),
-        ("image/webp", True),
-        ("image/avif", True),
-        ("application/json", False),
-        ("text/plain", False),
-        ("application/pdf", False),
-        (None, False),
-    ],
-)
-def test_is_valid_image_content_type(content_type, is_valid):
-    # when
-    result = is_valid_image_content_type(content_type)
-
-    # then
-    assert result == is_valid
 
 
 def test_clean_image_file():
@@ -272,23 +222,6 @@ def test_clean_image_file_in_avif_format():
 
     # when & then
     clean_image_file({field: img}, field, ProductErrorCode)
-
-
-@pytest.mark.parametrize(
-    ("content_type_header", "expected_mime_type"),
-    [
-        (None, None),
-        ("", ""),
-        ("text/html; charset=utf-8", "text/html"),
-        ("text/css; charset=UTF-8", "text/css"),
-        ("image/jpeg", "image/jpeg"),
-        ("  image/png; charset=binary", "image/png"),
-        (" Text/HTML ; Charset=UTF-8 ", "text/html"),
-        ("APPLICATION/JSON;CHARSET=UTF-8", "application/json"),
-    ],
-)
-def test_get_mime_type(content_type_header, expected_mime_type):
-    assert get_mime_type(content_type_header) == expected_mime_type
 
 
 @pytest.mark.parametrize(